+++ src/TicketCode/TicketCode.Core/Models/TcGroups.cs
@@ -28,18 +28,18 @@ namespace TicketCode.Core.Models
         /// 已使用的数量
         /// </summary>
         [Required]
-        public int iUsedNumber { get; set; }
+        public long iUsedNumber { get; set; }
 
         /// <summary>
         /// 每次递增的数量
         /// </summary>
         [Required]
-        public int iIncrNumber { get; set; }
+        public long iIncrNumber { get; set; }
 
         /// <summary>
         /// 低值预警
         /// </summary>
-        public int iMinNumber { get; set; }
+        public long iMinNumber { get; set; }
 
         [Required]
         public DateTime tCreateTime { get; set; }
@@ -48,7 +48,7 @@ namespace TicketCode.Core.Models
         /// 当前可用数量
         /// </summary>
         [Required]
-        public int iCurrAvaNumber { get; set; }
+        public long iCurrAvaNumber { get; set; }
 
         /// <summary>
         /// 更新时间
